pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "metis-app"
version = "0.1.0"
description = "METIS AI — Local-first frontier AI workspace with RAG, knowledge graphs, and an always-on companion"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "mrzapa" },
]
keywords = ["rag", "llm", "desktop", "pyside6", "qt", "cli"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
# Runtime dependencies are intentionally minimal; heavy ML libs (LangChain,
# sentence-transformers, llama-cpp-python) are optional/user-installed unless
# the installer opts into the pinned runtime-all bundle.
dependencies = ["PySide6>=6.6", "psutil>=6", "PyYAML>=6"]

[project.optional-dependencies]
api = [
    "fastapi==0.135.1",
    "python-multipart==0.0.22",
    "uvicorn[standard]==0.42.0",
]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.6",
    "pre-commit>=3.8",
    "httpx>=0.28",
    "fastapi==0.135.1",
    "python-multipart==0.0.22",
    "uvicorn[standard]==0.42.0",
    "litestar>=2.4",
]
runtime-all = [
    "langchain==0.3.28",
    "langchain-core==0.3.83",
    "langchain-text-splitters==0.3.11",
    "langchain-openai==0.3.35",
    "langchain-anthropic==0.3.22",
    "langchain-google-genai==2.1.12",
    "langchain-cohere==0.3.5",
    "langchain-voyageai==0.1.3",
    "langchain-community==0.3.31",
    "langchain-chroma==0.2.6",
    "langchain-weaviate==0.0.6",
    "chromadb>=0.5.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.7.0",
    "weaviate-client>=4.0.0",
    "sentence-transformers",
    "llama-cpp-python",
]
live-backends = [
    "weaviate-client>=4,<5",
]
kreuzberg = [
    "kreuzberg",
]
heretic = [
    "heretic-llm>=1.2",
]
all = [
    "kreuzberg",
    "weaviate-client>=4,<5",
]

[project.scripts]
metis-parity-audit = "metis_app.audit:main"
metis-api = "metis_app.api.__main__:main"

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
# Only package metis_app; exclude the legacy monolith and test tree.
include = ["metis_app*"]

[tool.setuptools.package-data]
"metis_app.assets" = ["logo.png", "llmfit_gguf_catalog.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-q --tb=short"

[tool.ruff]
exclude = [
  "agentic_rag_gui.py",
]
